decouple worker controller and launch script, write controller tests
