+++ main/signers/hot/SeedSigner/worker.ts
@@ -7,7 +7,7 @@ import {
   signMessage,
   signTransaction,
   signTypedData
-} from '../HotSigner/worker'
+} from '../HotSigner/controller'
 
 import type {
   CoreWorkerMethod,
