+++ test/main/signers/hot/RingSigner/worker.test.js
